+++ city_metrix/metrix_model.py
@@ -895,3 +895,20 @@ class Metric():
     def _verify_extension(file_path, extension):
         if Path(file_path).suffix != extension:
             raise ValueError(f"File name must have '{extension}' extension")
+
+
+def get_class_default_spatial_resolution(obj):
+    obj_param_info = get_param_info(obj.get_data)
+    obj_spatial_resolution = obj_param_info.get('spatial_resolution')
+    return obj_spatial_resolution
+
+
+def get_param_info(func):
+    import inspect
+    signature = inspect.signature(func)
+    default_values = {
+        k: v.default
+        for k, v in signature.parameters.items()
+        if v.default is not inspect.Parameter.empty
+    }
+    return default_values
