+++ tests/test_layer_parameters.py
@@ -4,9 +4,8 @@ import numpy as np
 from skimage.metrics import structural_similarity as ssim
 from pyproj import CRS
 from city_metrix.layers import *
-from city_metrix.metrix_model import GeoExtent, Layer
+from city_metrix.metrix_model import GeoExtent, Layer, get_class_default_spatial_resolution
 from tests.resources.bbox_constants import BBOX_BRA_LAURO_DE_FREITAS_1
-from tests.tools.general_tools import get_class_default_spatial_resolution
 from city_metrix.metrix_tools import get_class_from_instance
 
 COUNTRY_CODE_FOR_BBOX = 'BRA'
